openpilot v0.9.8 release
date: 2025-03-15T21:10:51
master commit: fb7b9c0f9420d228f03362970ebcfb7237095cf3

+++ conftest.py
@@ -8,6 +8,18 @@ from openpilot.common.prefix import OpenpilotPrefix
 from openpilot.system.manager import manager
 from openpilot.system.hardware import TICI, HARDWARE
 
+# TODO: pytest-cpp doesn't support FAIL, and we need to create test translations in sessionstart
+# pending https://github.com/pytest-dev/pytest-cpp/pull/147
+collect_ignore = [
+  "selfdrive/ui/tests/test_translations",
+  "selfdrive/test/process_replay/test_processes.py",
+  "selfdrive/test/process_replay/test_regen.py",
+]
+collect_ignore_glob = [
+  "selfdrive/debug/*.py",
+  "selfdrive/modeld/*.py",
+]
+
 
 def pytest_sessionstart(session):
   # TODO: fix tests and enable test order randomization
@@ -65,8 +77,10 @@ def openpilot_class_fixture():
 
 
 @pytest.fixture(scope="function")
-def tici_setup_fixture(openpilot_function_fixture):
+def tici_setup_fixture(request, openpilot_function_fixture):
   """Ensure a consistent state for tests on-device. Needs the openpilot function fixture to run first."""
+  if 'skip_tici_setup' in request.keywords:
+    return
   HARDWARE.initialize_hardware()
   HARDWARE.set_power_save(False)
   os.system("pkill -9 -f athena")
